Replace uint with double in Playlistitem sequencenumber

+++ server/src/TickTick.Api/Dtos/Songs/SongDetailDto.cs
@@ -5,6 +5,6 @@ public record SongDetailDto
         public string Song { get; set; }
         public string Title { get; set; }
         public TimeSpan? Duration { get; set; }
-        public uint SequenceNumber { get; set; }
+        public double SequenceNumber { get; set; }
 
 }
