+++ server/src/TickTick.Api/Dtos/Speeches/SpeechDetailDto.cs
@@ -4,7 +4,7 @@ public record SpeechDetailDto
 {
     public string Title { get; set; }
     public TimeSpan? Duration { get; set; }
-    public uint SequenceNumber { get; set; }
+    public double SequenceNumber { get; set; }
     public string? Speaker { get; set; }
     public string Text { get; set; }
 }
